requirements.txt
Flask==3.0.0
zeep==4.2.1
Werkzeug==3.0.1
anthropic
python-dotenv
gunicorn==21.2.0